git-svn-id: https://subversion.assembla.com/svn/kbasm/cpgf/trunk@644 5ea4cdcf-77fc-412b-903d-47c6f9a32d48

+++ test/variant/test_variant_cast.cpp
@@ -117,7 +117,7 @@ GTEST(TestVariant_CastFromFloat)
 
 	GCHECK(value.getType() == (byReference | vtFloat));
 
-	casted = fromVariant<float>(value);
+	casted = fromVariant<double>(value);
 	GCHECK(casted > 4.9f && casted < 5.1f);
 }
